Fix Backend

+++ app/Http/Controllers/RoomController.php
@@ -15,9 +15,25 @@ use Inertia\Inertia;
 class RoomController extends Controller
 {
     public function index(){
-        $rooms = Room::latest()->get();
+        $rooms = Room::latest()->get()->map(function($record){
+            $array = [];
+            $array["floor"]=["type"=>"text","data"=>$record->floor];
+            $array["number"]=["type"=>"text","data"=>$record->number];
 
-        return Inertia::render("room/index",compact("rooms"));
+            $array["button"]    = [ "type"=>"button",
+                "items"=>[
+                    ["data"=>route('room_show', $record->id)     ,  "method"=>"get"      ,"value"=>"نمایش"           , "type"=>"show"        ],
+                    ["data"=>route('device_room_create', $record->id)     ,  "method"=>"get"      ,"value"=>"افزودن وسیله"           , "type"=>"add"        ],
+                    ["data"=>route('room_edit', $record->id)     ,  "method"=>"get"      ,"value"=>"ویرایش"           , "type"=>"edit"        ],
+                    
+                    ["data"=>route("room_delete",$record->id)   ,  "method"=>"delete"   ,"value"=>"حذف"             , "type"=>"delete"      ],
+                ],
+            ];
+
+            return $array;
+        });
+        $header = ["طبقه","شماره","عملیات"];
+        return Inertia::render("room/index",compact("rooms","header"));
     }
     public function create(){
         return Inertia::render("room/create");
@@ -47,7 +63,9 @@ class RoomController extends Controller
 
     public function delete($id){
         $room = Room::findOrFail($id);
-        unlink(storage_path('app/public/pics/'.$room->picture));
+        if($room->picture){
+            unlink(storage_path('app/public/pics/'.$room->picture));
+        }
         $room->delete();
     }
